+++ panel-domain/src/main/java/com/emlakjet/advert/usecase/UpdateAdvertUseCase.java
@@ -6,7 +6,7 @@ import com.emlakjet.advert.model.LocationPoint;
 import com.emlakjet.commons.model.UseCase;
 import lombok.Builder;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Builder(toBuilder = true)
 public record UpdateAdvertUseCase(
@@ -14,7 +14,7 @@ public record UpdateAdvertUseCase(
         Long ownerId,
         String title,
         String description,
-        BigDecimal price,
+        BigInteger price,
         LocationPoint location,
         TradeType tradeType,
         IndoorInfo indoorInfo
